@@ -1,8 +1,7 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.zeeshanhussain.timetable"
         minSdkVersion 15
@@ -19,20 +18,20 @@ android {
 }
 
 dependencies {
-    compile 'com.roughike:bottom-bar:2.0.2'
-    compile 'com.ncapdevi:frag-nav:1.2.5'
-    compile 'com.shawnlin:number-picker:2.4.4'
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    compile 'com.daimajia.numberprogressbar:library:1.4@aar'
-    compile 'com.android.support:appcompat-v7:27.0.2'
-    compile 'com.android.support:design:27.0.2'
-    testCompile 'junit:junit:4.12'
-    compile 'com.android.support:support-v4:27.0.2'
-    compile 'com.android.support:recyclerview-v7:27.0.2'
-    compile "android.arch.persistence.room:runtime:1.0.0"
-    annotationProcessor "android.arch.persistence.room:compiler:1.0.0"
-    implementation "android.arch.lifecycle:extensions:1.1.0"
-    annotationProcessor "android.arch.lifecycle:compiler:1.1.0"
+    implementation 'com.roughike:bottom-bar:2.0.2'
+    implementation 'com.ncapdevi:frag-nav:1.2.5'
+    implementation 'com.shawnlin:number-picker:2.4.4'
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    implementation 'com.daimajia.numberprogressbar:library:1.4@aar'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    testImplementation 'junit:junit:4.12'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
+    implementation "androidx.room:room-runtime:2.0.0"
+    annotationProcessor "androidx.room:room-compiler:2.0.0"
+    implementation "androidx.lifecycle:lifecycle-extensions:2.0.0"
+    annotationProcessor "androidx.lifecycle:lifecycle-compiler:2.0.0"
     debugImplementation 'com.amitshekhar.android:debug-db:1.0.3'
 
 }
