@@ -1,11 +1,11 @@
 package com.zeeshanhussain.timetable.ui.activity;
 
-import android.support.v7.app.AppCompatActivity;
-import android.support.v4.app.Fragment;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.annotation.IdRes;
+import androidx.annotation.IdRes;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.widget.Toast;
