@@ -1,10 +1,10 @@
 package com.zeeshanhussain.timetable.viewmodel;
 
 import android.app.Application;
-import android.arch.lifecycle.AndroidViewModel;
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.MutableLiveData;
-import android.support.annotation.NonNull;
+import androidx.lifecycle.AndroidViewModel;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.annotation.NonNull;
 
 import com.zeeshanhussain.timetable.database.AppDatabase;
 import com.zeeshanhussain.timetable.model.Subject;
