@@ -1,10 +1,9 @@
 package com.zeeshanhussain.timetable.database;
 
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.Room;
-import android.arch.persistence.room.RoomDatabase;
+import androidx.room.Database;
+import androidx.room.Room;
+import androidx.room.RoomDatabase;
 import android.content.Context;
-import android.util.Log;
 
 import com.zeeshanhussain.timetable.model.Lecture;
 import com.zeeshanhussain.timetable.model.Subject;
