@@ -1,7 +1,7 @@
 package com.zeeshanhussain.timetable.viewmodel;
 
-import android.arch.lifecycle.ViewModel;
-import android.arch.lifecycle.ViewModelProvider;
+import androidx.lifecycle.ViewModel;
+import androidx.lifecycle.ViewModelProvider;
 
 import com.zeeshanhussain.timetable.database.AppDatabase;
 
